Move SelloDigitalCertificado to a trait with tests
