+++ tests/CfdiUtilsTests/Validate/Cfdi40/Standard/SelloDigitalCertificadoWithCfdiRegistroFiscalTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace CfdiUtilsTests\Validate\Cfdi33\Standard;
+namespace CfdiUtilsTests\Validate\Cfdi40\Standard;
 
-use CfdiUtils\Validate\Cfdi33\Standard\SelloDigitalCertificado;
+use CfdiUtils\Validate\Cfdi40\Standard\SelloDigitalCertificado;
 use CfdiUtilsTests\Validate\Common\SelloDigitalCertificadoWithCfdiRegistroFiscalTrait;
-use CfdiUtilsTests\Validate\Validate33TestCase;
+use CfdiUtilsTests\Validate\Validate40TestCase;
 
-final class SelloDigitalCertificadoWithCfdiRegistroFiscalTest extends Validate33TestCase
+final class SelloDigitalCertificadoWithCfdiRegistroFiscalTest extends Validate40TestCase
 {
     use SelloDigitalCertificadoWithCfdiRegistroFiscalTrait;
